another try at fixing tests
